transactional #finish

+++ mysite03/src/main/java/mysite/service/GuestbookService.java
@@ -12,8 +12,8 @@ import mysite.vo.GuestbookVo;
 @Service
 public class GuestbookService {
 	
-	private GuestbookRepository guestbookRepository;
-	private GuestbookLogRepository guestbookLogRepository;
+	private final GuestbookRepository guestbookRepository;
+	private final GuestbookLogRepository guestbookLogRepository;
 	
 	public GuestbookService(GuestbookRepository guestbookRepository, GuestbookLogRepository guestbookLogRepository) {
 		this.guestbookRepository = guestbookRepository;
@@ -21,16 +21,31 @@ public class GuestbookService {
 	}
 	
 	public List<GuestbookVo> getContentsList() {
-		List<GuestbookVo> list = guestbookRepository.findAll();
-		return list;
+		return guestbookRepository.findAll();
 	}
 	
 	@Transactional
 	public void deleteContents(Long id, String password) {
-		guestbookRepository.deleteByIdAndPassword(id, password);
+		GuestbookVo vo = guestbookRepository.findById(id);
+		if (vo == null) {
+			return;
+		}
+		
+		int count = guestbookRepository.deleteByIdAndPassword(id, password);
+		
+		if (count == 1) {
+			guestbookLogRepository.updateByRegDate(vo.getRegDate());
+		}
 	}
 	
+	@Transactional
 	public void addContents(GuestbookVo vo) {
+		int count = guestbookLogRepository.update();
+		
+		if (count == 0) { // 첫 글인 경우 
+			guestbookLogRepository.insert();
+		}
+		
 		guestbookRepository.insert(vo);
 	}
 }
